+++ documentation/ChangeLog.md
@@ -24,9 +24,16 @@
  * Fixed #46  0.10.0-PR adds spaces between characters to DUB extra options
  * DUB builder: incremental builds are now retried if previous 'dub build' exited with non-zero return code.
  * Fixed exception when dub.json modified and Package Explorer view was closed.
+ * Minor improvements to DUB elements in Package Explorer (added source folders for deps)
  TODO:
+   * correct getParent and equals
    * DLTK Model elements in Project Explorer
+     * flatten root script folder
+     * respond to DLTK model updates
+     * review Script label provider (text+image) 
    * cleanup loggers output
+   * fix buildpath update on project Add
+   * optimize performance issues with all projects DUB buildpath update
 
 ### DDT 0.9.1 (2014-02-07)
  * Added debug hover to D editor. When a CDT debug session is active, this hover presents detailed info for the variable under the cursor, similar to the Variables view.